refactoring namespaces
