+++ RehauSku.Assist.csproj
@@ -100,7 +100,7 @@
     <Compile Include="Source\Ribbon\RibbonController.cs" />
     <Compile Include="Source\Assistant\HttpClientUtil.cs" />
     <Compile Include="Source\Assistant\StoreResponse.cs" />
-    <Compile Include="Source\DataExport\DataWriter.cs" />
+    <Compile Include="Source\DataExport\Exporter.cs" />
     <Compile Include="Source\AddIn\AddIn.cs" />
     <Compile Include="Source\Assistant\IProduct.cs" />
     <Compile Include="Source\AddIn\Functions.cs" />
